backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "CloudMosa Backend API - A FastAPI-based backend server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.2",
    "pydantic-settings>=2.0.0",
]
